Move getExtensionStatus into communication
