+++ packages/bierzo-wallet/src/communication/extension.unit.spec.ts
